@@ -4,23 +4,21 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
-    "@stripe/react-stripe-js": "^1.1.2",
-    "@stripe/stripe-js": "^1.9.0",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "animated-scroll-to": "^2.0.10",
-    "axios": "^0.20.0",
-    "firebase": "^7.21.0",
-    "moment": "^2.29.0",
-    "react": "^16.13.1",
-    "react-currency-format": "^1.0.0",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3",
-    "stripe": "^8.100.0"
+    "@stripe/react-stripe-js": "^5.4.1",
+    "@stripe/stripe-js": "^8.5.3",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "animated-scroll-to": "^2.3.2",
+    "axios": "^1.13.2",
+    "firebase": "^12.6.0",
+    "moment": "^2.30.1",
+    "react": "^19.2.3",
+    "react-dom": "^19.2.3",
+    "react-number-format": "^5.4.4",
+    "react-router-dom": "^7.10.1",
+    "react-scripts": "^5.0.1",
+    "stripe": "^20.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -31,6 +29,13 @@
   },
   "eslintConfig": {
     "extends": "react-app"
+  },
+  "devDependencies": {
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
+    "@mui/icons-material": "^7.3.6",
+    "@mui/material": "^7.3.6",
+    "gh-pages": "^6.3.0"
   },
   "browserslist": {
     "production": [
@@ -43,8 +48,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "gh-pages": "^3.1.0"
   }
 }
