@@ -1,11 +1,11 @@
 import React from "react";
 import "../CSS/Header.css";
-import SearchIcon from "@material-ui/icons/Search";
-import ShoppingBasketIcon from "@material-ui/icons/ShoppingBasket";
+import SearchIcon from "@mui/icons-material/Search";
+import ShoppingBasketIcon from "@mui/icons-material/ShoppingBasket";
 import { Link } from "react-router-dom";
 import { useStateValue } from "../StateProvider";
 import { auth } from "../firebase";
-import MenuOpenIcon from '@material-ui/icons/MenuOpen';
+import MenuOpenIcon from '@mui/icons-material/MenuOpen';
 
 function Header() {
   const [{ basket, user }, dispatch] = useStateValue();
